New submitter submodule
