+++ panel/src/panel/modal.js
@@ -1,6 +1,5 @@
 // @ts-check
 
-import { isObject } from "@/helpers/object.js";
 import Feature, { defaults as featureDefaults } from "./feature.js";
 import focus from "@/helpers/focus.js";
 import "@/helpers/array.js";
@@ -168,158 +167,6 @@ export default (panel, key, defaults) => {
 			this.isOpen = true;
 
 			return this.state();
-		},
-
-		/**
-		 * Custom submitter for the dialog/drawer
-		 * It will automatically close the modal
-		 * if there's no submit listner or backend route.
-		 *
-		 * @param {Object} value
-		 * @param {Object} options
-		 * @returns {Promise} The new state or false if the request failed
-		 */
-		async submit(value, options = {}) {
-			// close the drawer or dialog if there's no submitter
-			// An example for this is the blocks drawer
-			if (this.hasSubmitter() === false) {
-				console.warn(`There's no ${this.key()} submitter`);
-				return this.close();
-			}
-
-			let response;
-
-			if (this.hasEventListener("submit")) {
-				// call a custom submit handler if it exists
-				response = await this.emit("submit", value ?? this.value, options);
-			} else {
-				// send a request to the backend
-				response = await this.post(value ?? this.value, options);
-			}
-
-			// the request failed and should have raised an error
-			if (isObject(response) === false) {
-				return response;
-			}
-
-			// get details from the response object.
-			// I.e. { $dialog: { ... } }
-			// pass it forward to the success handler
-			// to react on elements in the response
-			return this.success(response["$" + this.key()] ?? {});
-		},
-
-		/**
-		 * This is rebuilding the previous
-		 * behaviours from the dialog mixin.
-		 * Most of the response handling should
-		 * be redone. But we keep it for compatibility
-		 *
-		 * @param {Object|String} success
-		 * @returns
-		 */
-		success(success) {
-			if (typeof success === "string") {
-				panel.notification.success(success);
-			}
-
-			// close the modal
-			this.close();
-
-			// show a success message
-			this.successNotification(success);
-
-			// send custom events to the event bus
-			this.successEvents(success);
-
-			// dispatch store actions that might have been defined in the response
-			this.successDispatch(success);
-
-			// redirect or reload
-			if (success.route || success.redirect) {
-				// handle any redirects
-				this.successRedirect(success);
-			} else {
-				// reload the parent view to show changes
-				panel.view.reload(success.reload);
-			}
-
-			return success;
-		},
-
-		/**
-		 * Dispatch deprecated store events
-		 *
-		 * @param {Object} state
-		 */
-		successDispatch(state) {
-			if (isObject(state.dispatch) === false) {
-				return;
-			}
-
-			for (const event in state.dispatch) {
-				const payload = state.dispatch[event];
-				panel.app.$store.dispatch(
-					event,
-					Array.isArray(payload) === true ? [...payload] : payload
-				);
-			}
-		},
-
-		/**
-		 * Emit all events that might be in the response
-		 *
-		 * @param {Object} state
-		 */
-		successEvents(state) {
-			if (state.event) {
-				// wrap single events to treat them all at once
-				const events = Array.wrap(state.event);
-
-				// emit all defined events
-				for (const event of events) {
-					if (typeof event === "string") {
-						panel.events.emit(event, state);
-					}
-				}
-			}
-
-			// emit a success event
-			if (state.emit !== false) {
-				panel.events.emit("success", state);
-			}
-		},
-
-		/**
-		 * Sends a success notification if the
-		 * response contains a success message
-		 *
-		 * @param {Object} state
-		 */
-		successNotification(state) {
-			if (state.message) {
-				panel.notification.success(state.message);
-			}
-		},
-
-		/**
-		 * Handles redirects in submit responses
-		 *
-		 * @param {Object} state
-		 */
-		successRedirect(state) {
-			const redirect = state.route ?? state.redirect;
-
-			// no redirect
-			if (!redirect) {
-				return false;
-			}
-
-			if (typeof redirect === "string") {
-				return panel.open(redirect);
-			}
-
-			return panel.open(redirect.url, redirect.options);
 		},
 
 		/**
